requirements.txt
# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Image Processing
Pillow>=9.5.0

# AI/ML Models
torch>=2.0.0
transformers>=4.30.0
einops>=0.6.1
timm>=0.9.2

# Web Framework
streamlit>=1.25.0

# Financial Data
yfinance>=0.2.18

# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0

# Database
sqlite3

# Utilities
glob2>=0.7
os-sys>=2.1.4

# Optional: For better performance
# accelerate>=0.20.0
# huggingface_hub[hf_xet]>=0.15.0
